+++ js/constans.js
@@ -1,0 +1,29 @@
+export const DESCRIPTION = [
+  'Утро',
+  'Котик',
+  'Солнышко',
+  'Море',
+  'Облака',
+];
+
+export const NAME = [
+  'Иван',
+  'Елена',
+  'Игорь',
+  'Анна',
+];
+
+export const MESSAGE = [
+  'Всё отлично!',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+];
+
+export const PHOTOS_COUNT = 25;
+export const MIN_ID_COMMENT = 2000;
+export const MAX_ID_COMMENT = 10000;
+export const MIN_AVATAR = 1;
+export const MAX_AVATAR = 6;
+export const MIN_COMMENTS = 0;
+export const MAX_COMMENTS = 30;
+export const MIN_LIKES = 15;
+export const MAX_LIKES = 200;
